organize code to be testable in python 3.5
